+++ pubspec.yaml
@@ -37,7 +37,6 @@ dependencies:
   flutter_svg: ^0.22.0
   fluttertoast: ^8.0.6
   flutter_slidable: ^0.6.0
-  firebase_messaging: ^9.1.3
 
 dev_dependencies:
   flutter_test:
